refactor(sp): Move `assign_proving_period_offset` to primitives
