+++ pallets/storage-provider/src/proofs.rs
@@ -8,7 +8,6 @@ use primitives::{
     MAX_POST_PROOFS_PER_BLOCK, MAX_POST_PROOF_BYTES,
 };
 use scale_info::TypeInfo;
-use sp_core::blake2_64;
 
 /// Proof of Spacetime data stored on chain.
 #[derive(RuntimeDebug, Decode, Encode, TypeInfo, PartialEq, Eq, Clone)]
@@ -31,41 +30,4 @@ pub struct SubmitWindowedPoStParams {
     pub partitions: BoundedVec<PartitionNumber, ConstU32<MAX_PARTITIONS_PER_DEADLINE>>,
     /// The proof submission.
     pub proofs: BoundedVec<PoStProof, ConstU32<MAX_POST_PROOFS_PER_BLOCK>>,
-}
-
-/// Error type for proof operations.
-#[derive(RuntimeDebug)]
-pub enum ProofError {
-    Conversion,
-}
-
-/// Assigns proving period offset randomly in the range [0, WPOST_PROVING_PERIOD)
-/// by hashing the address and current block number.
-///
-/// Reference:
-/// * <https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/miner/src/lib.rs#L4886>
-pub(crate) fn assign_proving_period_offset<AccountId, BlockNumber>(
-    addr: &AccountId,
-    current_block: BlockNumber,
-    wpost_proving_period: BlockNumber,
-) -> Result<BlockNumber, ProofError>
-where
-    AccountId: Encode,
-    BlockNumber: sp_runtime::traits::BlockNumber,
-{
-    // Encode address and current block number
-    let mut addr = addr.encode();
-    let mut block_num = current_block.encode();
-    // Concatenate the encoded block number to the encoded address.
-    addr.append(&mut block_num);
-    // Hash the address and current block number for a pseudo-random offset.
-    let digest = blake2_64(&addr);
-    // Create a pseudo-random offset from the bytes of the hash of the address and current block number.
-    let offset = u64::from_be_bytes(digest);
-    // Convert into block number
-    let mut offset =
-        TryInto::<BlockNumber>::try_into(offset).map_err(|_| ProofError::Conversion)?;
-    // Mod with the proving period so it is within the valid range of [0, WPOST_PROVING_PERIOD)
-    offset %= wpost_proving_period;
-    Ok(offset)
 }
